refactor: move input group display component to components direcotry
