+++ src/app/components/input-group-display/input-group-display.component.spec.ts
